feat(processor.go): dynamically determine ID column name based on table name
This change modifies the generateSelectStatement function to dynamically
determine the ID column name based on the table name. It introduces a
switch statement that checks the lowercase version of the table name and
assigns the appropriate ID column name. The default case constructs the
ID column name by appending "_ID" to the table name. This change ensures
that the correct ID column is used in the SELECT statement, regardless
of the table being queried.